fix(e2e): extract shared PII helper, fix broken isVisible pattern + products readiness
- Extract dismissPIIModalIfPresent to e2e/utils/pii-helpers.ts (shared)
- Replace broken isVisible({timeout:3000}) in auth.spec.ts with shared helper
- Fix products.spec.ts: header → getByTestId('dashboard-button') for auth readiness

Addresses quality gate findings: auth.spec.ts had the same broken PII
detection pattern that maximotodev flagged in fixtures, and products.spec.ts
still used generic header check instead of authenticated-only marker.
